+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11RequestParser.java
@@ -1,17 +1,19 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
+import org.apache.coyote.http11.Cookie;
+import org.apache.coyote.http11.Http11Header;
 
-public class Http11RequestParser {
+class Http11RequestParser {
 
     /*
     https://www.rfc-editor.org/rfc/rfc2616#section-5 이 문서에서 이 메서드가 반환하는 것을 Request-URI라 지칭합니다.
      */
-    public String parseRequestURI(String requestMessage) {
+    String parseRequestURI(String requestMessage) {
         String startLine = parseStartLine(requestMessage);
         validateStartLine(startLine);
         return startLine.split(" ")[1];
@@ -28,7 +30,7 @@ public class Http11RequestParser {
         return requestMessage.split("\r\n")[0];
     }
 
-    public String readAsString(InputStream inputStream) {
+    String readAsString(InputStream inputStream) {
         try {
             // 학습 InputStream 의 readAllBytes 는 요청 전체를 읽는다. 이때, EOF를 만날때 까지 읽으므로 EOF가 없다면 Integer.MAX_VALUE 크기만큼 읽는다.
             int readByte = inputStream.available();
@@ -39,13 +41,13 @@ public class Http11RequestParser {
         }
     }
 
-    public Http11Method parseMethod(String requestMessage) {
+    Http11Method parseMethod(String requestMessage) {
         String startLine = parseStartLine(requestMessage);
         String rawMethod = startLine.split(" ")[0];
         return Http11Method.valueOf(rawMethod.toUpperCase());
     }
 
-    public List<Http11Header> parseHeaders(String requestMessage) {
+    List<Http11Header> parseHeaders(String requestMessage) {
         String rawHeaders = requestMessage.split("\r\n\r\n")[0]
                 .replace(parseStartLine(requestMessage), "")
                 .replaceFirst("\r\n", "");
@@ -61,7 +63,7 @@ public class Http11RequestParser {
                 .toList();
     }
 
-    public List<Cookie> parseCookies(String requestMessage) {
+    List<Cookie> parseCookies(String requestMessage) {
         String rawHeaders = requestMessage.split("\r\n\r\n")[0]
                 .replace(parseStartLine(requestMessage), "")
                 .replaceFirst("\r\n", "");
@@ -84,7 +86,7 @@ public class Http11RequestParser {
         return new Cookie(key, value);
     }
 
-    public LinkedHashMap<String, String> parseBody(String requestMessage) {
+    LinkedHashMap<String, String> parseBody(String requestMessage) {
         int startIndex = requestMessage.indexOf("\r\n\r\n");
         if (startIndex == -1) {
             return new LinkedHashMap<>();
